feat: request context setup
